refactor: improving export statements
