+++ src/products/application/decotators/index.ts
@@ -1,0 +1,2 @@
+export * from './open-api.decorator';
+export * from './repository.decorator';
